chore: Move string-utils.js to utils folder
