+++ src/js/utils/string-utils.js
